Phase 5.7 & 5.8: Migrate validate() and clean() to canonical modules
Migrated validate() and clean() into _validate.py and _clean.py respectively. Clean() now uses sys.modules Rich resolution pattern, mirroring _merge.py and others. Replaced monolithic inline implementations in __init__.py with thin imports. Regression tests pass.
