+++ python/zedda/__init__.py
@@ -1069,367 +1069,7 @@ from ._fix import fix
 #  pandas, creates a backup file, and writes a JSON audit trail.
 #  Shows before/after quality scores with visual progress.
 # ─────────────────────────────────────────────────────────────────
-def clean(path, output: str | None = None, sample_size: int | None = None) -> Any:
-    """
-    Auto-clean a dataset by applying all auto-fixable warnings.
-
-    Creates a backup, applies fixes (impute, drop, encode), and
-    saves the cleaned file with a JSON audit trail.
-
-    Args:
-        path (str): Path to a ``.csv``, ``.parquet``, or ``.arrow`` file.
-        output (str, optional): Output file path. If None, overwrites
-            the original (after creating a backup).
-        sample_size (int, optional): Max rows to sample for profiling.
-
-    Returns:
-        pandas.DataFrame: The cleaned DataFrame.
-
-    Example::
-
-        import zedda as zd
-        zd.clean("titanic.csv", output="titanic_clean.csv")
-        zd.clean.undo("titanic.csv")  # restore from backup
-    """
-    # FIX L-22: json and shutil are already imported at module level via
-    # clean()'s body. These redundant imports add overhead on every call.
-    # They're now imported at the top of clean() only if not already available.
-
-    resolved_path, is_in_memory = _resolve_input(path)
-    try:
-        import pandas as pd
-    except ImportError:
-        raise ZeddaError("pandas is required for clean(). Run: pip install pandas")
-
-    try:
-        if not _RICH_AVAILABLE or _console is None:
-            raise ZeddaError(
-                "Rich is required for terminal output. Install with: pip install rich"
-            )
-
-        file_name = (
-            "<DataFrame>"
-            if is_in_memory
-            else (Path(path).name if isinstance(path, (str, Path)) else "<DataFrame>")
-        )
-
-        crit_sym = _safe_symbol("✗", "[X]")
-        warn_sym = _safe_symbol("⚠", "[!]")
-        check_sym = _safe_symbol("✓", "[OK]")
-        arrow_r = _safe_symbol("→", "->")
-        bullet = _safe_symbol("·", "-")
-
-        # ── Header ──────────────────────────────────────────────────
-        _console.print(
-            f"\n[bold blue]zedda[/bold blue] [dim]v{__version__}[/dim]  {bullet}  "
-            f"[bold]clean mode[/bold]\n"
-        )
-
-        # ── Profile BEFORE ──────────────────────────────────────────
-        p = _scan_wrapper(resolved_path, sample_size=sample_size)
-        all_warnings = _collect_warnings(p)
-        fixable = [w for w in all_warnings if w.get("auto_fixable")]
-
-        score_before = _quality_score(p)
-        n_critical = sum(1 for w in all_warnings if w["severity"] == "critical")
-        n_warning = sum(1 for w in all_warnings if w["severity"] == "warning")
-        n_info = sum(1 for w in all_warnings if w["severity"] == "info")
-
-        bar = _render_quality_bar(score_before)
-        color, label = _quality_label(score_before)
-
-        _console.print("[bold]Before[/bold]")
-        _console.print(
-            f"  Quality score : [{color}]{score_before}/100  {bar}  {label}[/{color}]"
-        )
-        _console.print(
-            f"  Issues found  : {len(all_warnings)}  "
-            f"({n_critical} critical {bullet} {n_warning} warning{'s' if n_warning != 1 else ''}"
-            f" {bullet} {n_info} info)\n"
-        )
-
-        if not fixable:
-            _console.print(
-                f"  [green]{check_sym}  No auto-fixable issues — data is already clean![/green]\n"
-            )
-            return None
-
-        # ── Load the data ───────────────────────────────────────────
-        if is_in_memory:
-            if "polars" in getattr(type(path), "__module__", ""):
-                df = path.to_pandas()
-            else:
-                df = path.copy()
-        else:
-            assert isinstance(resolved_path, str)
-            ext = Path(resolved_path).suffix.lower()
-            if ext == ".csv":
-                df = pd.read_csv(resolved_path)
-            elif ext in (".parquet", ".arrow"):
-                df = pd.read_parquet(resolved_path)
-            else:
-                raise ZeddaError(f"Unsupported format for clean: {ext}")
-
-        rows_before = len(df)
-        cols_before = len(df.columns)
-
-        # ── Create backup ───────────────────────────────────────────
-        if not is_in_memory:
-            assert isinstance(resolved_path, str)
-            backup_path = str(resolved_path) + ".zedda-backup"
-            shutil.copy2(resolved_path, backup_path)
-            _console.print("[bold]Backup[/bold]")
-            _console.print(
-                f"  [green]{check_sym}[/green]  Backup saved {arrow_r} {Path(backup_path).name}"
-            )
-            _console.print(f'     Restore anytime: zd.clean.undo("{file_name}")\n')
-        else:
-            backup_path = None
-
-        # ── Apply fixes ─────────────────────────────────────────────
-        _console.print("[bold]Applying Fixes[/bold]")
-        audit_actions = []
-        dropped_cols = []
-
-        for w in fixable:
-            col_name = w["column"]
-            action = w["action_type"]
-            safe_display = rich_escape(col_name)
-
-            if action == "drop":
-                if col_name in df.columns:
-                    reason = w["message"]
-                    df = df.drop(columns=[col_name])
-                    dropped_cols.append(col_name)
-                    _console.print(
-                        f"  [green]{check_sym}[/green]  {safe_display} {arrow_r} dropped ({reason})"
-                        f"      [dim]col removed[/dim]"
-                    )
-                    audit_actions.append(
-                        {
-                            "column": col_name,
-                            "action": "drop",
-                            "reason": w["message"],
-                        }
-                    )
-
-            elif action == "impute":
-                if col_name in df.columns:
-                    col_data = df[col_name]
-                    col_obj = next((c for c in p.columns if c.name == col_name), None)
-                    null_count = (
-                        int(col_obj.null_count)
-                        if col_obj
-                        else int(col_data.isnull().sum())
-                    )
-                    if col_obj and col_obj.type_str in ("int", "float"):
-                        coerced_data = pd.to_numeric(col_data, errors="coerce")
-                        coerced_count = max(
-                            0, int(coerced_data.isnull().sum() - null_count)
-                        )
-
-                        fill_val = coerced_data.median()
-                        df[col_name] = coerced_data.fillna(fill_val)
-
-                        _console.print(
-                            f"  [green]{check_sym}[/green]  {safe_display}"
-                            f" {arrow_r} median imputed ({fill_val:.2f})"
-                            f"      [dim]{null_count + coerced_count} cells[/dim]"
-                        )
-                        if coerced_count > 0:
-                            _console.print(
-                                f"     [yellow]{warn_sym}[/yellow]  {safe_display} — {coerced_count} values could not be parsed as numbers and were treated as missing before imputation."
-                            )
-                    else:
-                        m = col_data.mode()
-                        fill_val = m[0] if not m.empty else "Unknown"
-                        df[col_name] = col_data.fillna(fill_val)
-                        _console.print(
-                            f"  [green]{check_sym}[/green]  {safe_display}"
-                            f" {arrow_r} mode imputed ({fill_val})"
-                            f"      [dim]{null_count} cells[/dim]"
-                        )
-                    audit_actions.append(
-                        {
-                            "column": col_name,
-                            "action": "impute",
-                            "fill_value": str(fill_val),
-                            "cells_fixed": null_count,
-                        }
-                    )
-
-            elif action == "encode":
-                if col_name in df.columns:
-                    n_unique = df[col_name].nunique()
-                    df[col_name] = pd.Categorical(df[col_name]).codes
-                    _console.print(
-                        f"  [green]{check_sym}[/green]  {safe_display}"
-                        f" {arrow_r} label encoded ({n_unique} unique)"
-                        f"      [dim]encoded[/dim]"
-                    )
-                    audit_actions.append(
-                        {
-                            "column": col_name,
-                            "action": "encode",
-                            "unique_values": n_unique,
-                        }
-                    )
-
-            elif action == "clip":
-                if col_name in df.columns:
-                    upper = df[col_name].quantile(0.99)
-                    clipped = int((df[col_name] > upper).sum())
-                    df[col_name] = df[col_name].clip(upper=upper)
-                    _console.print(
-                        f"  [green]{check_sym}[/green]  {safe_display}"
-                        f" {arrow_r} clipped at p99 ({upper:.2f})"
-                        f"      [dim]{clipped} cells[/dim]"
-                    )
-                    audit_actions.append(
-                        {
-                            "column": col_name,
-                            "action": "clip",
-                            "upper_bound": float(upper),
-                            "cells_clipped": clipped,
-                        }
-                    )
-
-        # ── Compute AFTER score ─────────────────────────────────────
-        import tempfile
-
-        tmp = tempfile.NamedTemporaryFile(suffix=".parquet", delete=False)
-        tmp.close()
-        score_after: int | None = None
-        rescan_error: str | None = None
-        try:
-            _require_pyarrow()
-            import pyarrow as pa
-            import pyarrow.parquet as pq
-
-            table = pa.Table.from_pandas(df, preserve_index=False)
-            pq.write_table(table, tmp.name)
-            p_after = _scan_wrapper(tmp.name)
-            score_after = _quality_score(p_after, original_cols=cols_before)
-        except Exception as e:
-            rescan_error = str(e)
-        finally:
-            _cleanup_temp(tmp.name)
-
-        if score_after is None:
-            score_after = score_before
-
-        improvement = score_after - score_before
-        rows_after = len(df)
-        cols_after = len(df.columns)
-
-        bar_a = _render_quality_bar(score_after)
-        color_a, label_a = _quality_label(score_after)
-
-        _console.print("\n[bold]After[/bold]")
-        sign = "+" if improvement >= 0 else ""
-        color_imp = "green" if improvement >= 0 else "red"
-        _console.print(
-            f"  Quality score : [{color_a}]{score_after}/100  "
-            f"{bar_a}  {label_a}[/{color_a}]"
-            f"  [{color_imp}]({sign}{improvement} points)[/{color_imp}]"
-        )
-        n_dropped = len(dropped_cols)
-        _console.print(
-            f"  Rows : {rows_before:,} {arrow_r} {rows_after:,}   "
-            f"Cols : {cols_before} {arrow_r} {cols_after}"
-            + (f"  ({n_dropped} dropped)" if n_dropped > 0 else "")
-        )
-
-        # ── Save output ─────────────────────────────────────────────
-        t0 = time.perf_counter()
-        if is_in_memory and not output:
-            out_path = None
-            elapsed = 0.0
-        else:
-            out_path = output if output else str(resolved_path)
-            out_ext = Path(out_path).suffix.lower()
-            out_parent = Path(out_path).resolve().parent
-            if not out_parent.exists():
-                raise ZeddaError(f"Output directory does not exist: '{out_parent}'")
-            if out_ext in (".parquet", ".arrow"):
-                df.to_parquet(out_path, index=False)
-            else:
-                df.to_csv(out_path, index=False)
-            elapsed = (time.perf_counter() - t0) * 1000
-
-        # ── Audit trail ─────────────────────────────────────────────
-        audit_path = None
-        if out_path is not None:
-            audit_path = str(Path(out_path).with_suffix("")) + ".audit.json"
-            if Path(audit_path).resolve().parent != Path(out_path).resolve().parent:
-                raise ZeddaError("Audit path traversal detected — refusing to write.")
-            audit_data = {
-                "source_file": file_name,
-                "output_file": Path(out_path).name,
-                "zedda_version": __version__,
-                "score_before": score_before,
-                "score_after": score_after,
-                "rows_before": rows_before,
-                "rows_after": rows_after,
-                "cols_before": cols_before,
-                "cols_after": cols_after,
-                "actions": audit_actions,
-            }
-            with open(audit_path, "w", encoding="utf-8") as f:
-                json.dump(audit_data, f, indent=2, ensure_ascii=False)
-
-        _console.print("\n[bold]Output[/bold]")
-        if out_path is not None:
-            _console.print(
-                f"  [green]{check_sym}[/green]  Clean file  {arrow_r} {Path(out_path).name}"
-            )
-            if audit_path:
-                _console.print(
-                    f"  [green]{check_sym}[/green]  Audit trail {arrow_r} {Path(audit_path).name}"
-                )
-            if backup_path:
-                _console.print(
-                    f"     Time: {elapsed:.1f}ms  {bullet}  Backup: {Path(backup_path).name}\n"
-                )
-            else:
-                _console.print(f"     Time: {elapsed:.1f}ms\n")
-        else:
-            _console.print(
-                f"  [green]{check_sym}[/green]  Returned cleaned DataFrame "
-                "(no file written — input was a DataFrame and `output` was not set)."
-            )
-            if backup_path:
-                _console.print(f"     Backup: {Path(backup_path).name}\n")
-            else:
-                _console.print()
-
-        return df
-
-    finally:
-        if is_in_memory:
-            _cleanup_temp(resolved_path)
-
-
-def _clean_undo(path) -> None:
-    """Restore a file from its zedda backup."""
-    import shutil
-
-    check_sym = _safe_symbol("✓", "[OK]")
-    backup = str(path) + ".zedda-backup"
-    if not Path(backup).exists():
-        raise ZeddaError(
-            f"No backup found: '{backup}'\n"
-            "Tip: zd.clean() creates a backup before modifying files."
-        )
-    shutil.copy2(backup, str(path))
-    if _RICH_AVAILABLE and _console:
-        _console.print(
-            f"\n[green]{check_sym}[/green]  Restored [cyan]{Path(path).name}[/cyan] "
-            f"from backup.\n"
-        )
-    else:
-        print(f"Restored {path} from backup.")
-
+from ._clean import clean, _clean_undo
 
 # P-05: Attach undo as a callable attribute on clean.
 # Type checkers can't see monkey-patched attributes by default.
